+++ browser.js
@@ -18,8 +18,7 @@ var Module = {
   Grid: require('./lib/grid/Component'),
   Form: require('./lib/form/FormService'),
   connectForm: require('./lib/form/connectForm'),
-  createValidator: require('./lib/common/validation/validators/browser').create,
-  createXhrValidator: require('./lib/common/validation/validators/XhrValidator').create,
+  createValidator: require('./lib/common/validation/Validator').create,
   exportGridData: require('./lib/grid/export/exportGridData'),
   toJSON: require('./lib/grid/export/exporters/toJSON'),
   ValidationErrors: require('./lib/common/validation/ValidationErrors'),
@@ -31,7 +30,6 @@ var Module = {
     Events: require('./lib/common/Events'),
     Form: require('./lib/form/FormModel'),
     FormXhr: require('./lib/form/FormXhrModel'),
-    ValidationErrors: require('./lib/common/validation/ValidationErrors'), // Deprecated. Use UIKernel.ValidationErrors
     List: {
       Xhr: require('./lib/list/ListXhrModel')
     }