+++ main.js
@@ -11,7 +11,6 @@ var browserModule = require('./browser');
 browserModule.gridExpressApi = require('./lib/grid/models/GridExpressApi').create;
 browserModule.listExpressApi = require('./lib/list/ListExpressApi').create;
 browserModule.formExpressApi = require('./lib/form/FormExpressApi').create;
-browserModule.createValidator = require('./lib/common/validation/validators/common').create;
 browserModule.toCSV = require('./lib/grid/export/exporters/toCSV');
 
 module.exports = browserModule;