+++ src/form/adapters/GridToFormUpdate.js
@@ -125,7 +125,7 @@ class GridToFormUpdate extends Events {
    */
   async isValidRecord(record) {
     const model = this._adapter.model;
-    return await model.isValidRecord(record);
+    return await model.isValidRecord(record, this._adapter.id);
   }
 
   /**